Split editable device display name into component
